add auto layer, multiple other fixes
